Updated README.md
Put a little note in about conflicted copies. It'd be interesting to see
how git handles it, but for now we just need to set this whole thing up.

This is a small project that hardly 'warrants' source management, but
it's what the real world uses for larger projects. If we can master it
for small stuff, we'll be decent with the big stuff.